[script.kodi.loguploader] 1.0.2
