+++ script.kodi.loguploader/addon.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="script.kodi.loguploader" name="Kodi Logfile Uploader" version="1.0.1" provider-name="Team Kodi">
+<addon id="script.kodi.loguploader" name="Kodi Logfile Uploader" version="1.0.2" provider-name="Team Kodi">
 	<requires>
 		<import addon="xbmc.python" version="3.0.0"/>
-		<import addon="script.module.pyqrcode" version="1.2.1"/>
-		<import addon="script.module.requests" version="2.19.1"/>
+		<import addon="script.module.pyqrcode" version="1.2.1+matrix.1"/>
+		<import addon="script.module.requests" version="2.22.0+matrix.1"/>
 	</requires>
 	<extension point="xbmc.python.script" library="default.py">
 		<provides>executable</provides>
@@ -11,15 +11,13 @@
 	<extension point="xbmc.addon.metadata">
 		<summary lang="en_GB">Uploads your kodi.log file.</summary>
 		<description lang="en_GB">Uploads your kodi.log file and returns an url you can post on http://forum.kodi.tv/</description>
-		<language></language>
 		<platform>all</platform>
-		<license>GNU GENERAL PUBLIC LICENSE Version 2</license>
+		<license>GPL-2.0-only</license>
 		<forum>https://forum.kodi.tv/showthread.php?tid=288153</forum>
-		<website>https://kodi.tv/</website>
-		<email></email>
 		<source>https://gitlab.com/ronie/script.kodi.loguploader/</source>
 		<assets>
 			<icon>resources/icon.png</icon>
 		</assets>
+		<news>- code refactor</news>
 	</extension>
 </addon>
